# backend/internal/docs/swagger.yaml
definitions:
...
  handler.MarkReadRequest:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  handler.SendMessageRequest:
    properties:
      content:
        type: string
      to_id:
        type: integer
    required:
    - content
    - to_id
    type: object
...
  response.PageResponse:
    properties:
      has_more:
        type: boolean
      items: {}
      next_cursor:
        type: string
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: 高性能短视频 Feed 流系统 API 文档
  termsOfService: http://swagger.io/terms/
  title: TideFlow 短视频 Feed 流 API
  version: v1
paths:
...
  /api/v1/messages:
    post:
      consumes:
      - application/json
      description: 向指定用户发送私信
      parameters:
      - description: 私信内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.SendMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 发送私信
      tags:
      - 私信
  /api/v1/messages/conversations:
    get:
      description: 获取当前用户的会话列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 会话列表
      tags:
      - 私信
  /api/v1/messages/conversations/{user_id}:
    get:
      description: 获取与指定用户的聊天记录
      parameters:
      - description: 用户ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: 游标
        in: query
        name: cursor
        type: string
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 获取消息记录
      tags:
      - 私信
  /api/v1/messages/conversations/{user_id}/read:
    put:
      description: 标记与指定用户的会话为已读
      parameters:
      - description: 用户ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 标记会话已读
      tags:
      - 私信
  /api/v1/notifications:
    get:
      description: 获取当前用户的通知列表
      parameters:
      - description: 游标
        in: query
        name: cursor
        type: string
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 通知列表
      tags:
      - 通知
  /api/v1/notifications/read:
    put:
      consumes:
      - application/json
      description: 标记通知为已读，传入IDs只标记指定通知，不传或空数组标记全部
      parameters:
      - description: 通知ID列表
        in: body
        name: body
        schema:
          $ref: '#/definitions/handler.MarkReadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 标记通知已读
      tags:
      - 通知
  /api/v1/notifications/stream:
    get:
      description: 建立SSE连接，接收实时通知推送
      parameters:
      - description: 访问令牌
        in: query
        name: token
        required: true
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE事件流
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: SSE实时通知
      tags:
      - 通知
  /api/v1/notifications/unread-count:
    get:
      description: 获取当前用户未读通知的数量
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 未读通知数
      tags:
      - 通知
...
securityDefinitions:
  OAuth2Password:
    flow: password
    scopes:
      read: Grant read access
      write: Grant write access
    tokenUrl: /api/v1/auth/oauth2/token
    type: oauth2
swagger: "2.0"